+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [0.14.0]  2022-07-20
 - Add Insert Path support
   - insert, into, values, and returning steps
+  - Values step can take binders
 
 ## [0.13.1]  2022-07-20
 ### Added